float added to sub

+++ calc.c
@@ -2,7 +2,7 @@
 double add(double x,double y){
 	return x+y;
 }
-int sub(int x,int y){
+float sub(float x,float y){
 	return x-y;
 }
 double mul(double x,double y){
